Server: Check admin is not null in Company model
This change is because I added a company by default in the database,
useful to have certain clients that do not belong to a specific company.
This company does not have an admin, so the Company model would fail when
trying to convert the admin (user) property to an array.