Client: Improve some form components
Add the disabled property to form-field, with the specific goal of using it in checkbox-group and checkbox.
Add the disabled property to submit-button.
Now Form component can work with a partial values object (values={{justOneProp: justOnePropVal}}). Before this improvement, it was mandatory to pass an object with the values of all the fields, otherwise the missing fields would not retain the entered data (except for the select and checkbox types).