Client: Check all departments, and disable the checkbox group, when staff level 3 (administrator) is selected.
This is informative and does not have a real effect when saving the changes, because on the server, if the staff level is 3, the staff is automatically assigned to all departments.
Remember that staff level 3 is the maximum role in the system, it is an administrator, and must be able to access all the information, and before this change, he could not see the tickets of the departments to which he did not belong.